ai-service/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
google-genai>=1,<2
pydantic>=2,<3
python-dotenv>=1,<2

